Run even more on cloud
This new version runs the most we can on cloud.
The bano dataset now lives on cloud and the person indices as well.

Which means that we don't have anymore Elasticsearch or Kibana running locally.

Only filebeat and logstash are running locally.

Filebeat reads from an http service or from the csv files and send the documents to an ingest pipeline.
Logstash reads and writes from and to Elasticsearch in cloud.